settings: add keep-alive options
In a server/client architecture, the client can loose its connectivity.
In this configuration, if the socket is not used, it will remains opened
unti the server try to use it. In some cases, there is nothing to send to
the client. The server should be able to send KEEP-ALIVE packets to
detect unreachable clients. This patch adds some settings to configure
this keep alive mechanism.

Issue: #324